Move some code files to the specific folder
